+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Kotlin
   * Change Metrics Ping Scheduler to use daemon threads ([#2930](https://github.com/mozilla/glean/pull/2930))
+  * Dispatch metric recording for event, object and timing distribution on the task queue ([#2942](https://github.com/mozilla/glean/pull/2942))
 
 # v61.0.0 (2024-08-21)